+++ src/components/proposal/table/row/index.ts
@@ -1,0 +1,2 @@
+export { ProposalItemActionsMenu } from './ProposalItemActionsMenu';
+export { ProposalRow } from './ProposalRow';